+++ app/src/main/java/com/masthuggis/boki/view/ThumbnailView.java
@@ -3,7 +3,7 @@ import com.masthuggis.boki.model.Advert;
 
 public interface ThumbnailView {
     void setTitle(String name);
-    void setPrice(int price);
+    void setPrice(long price);
     void setImageUrl(String url);
     void setId(String id);
     void setCondition(Advert.Condition conditon);
